# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: NutriAI — AI Nutrition Assistant API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /meals:
    get:
      operationId: listMeals
      tags: [meals]
      summary: List meals for the current user
      security:
        - bearerAuth: []
      parameters:
        - name: date
          in: query
          description: Filter by date (YYYY-MM-DD)
          schema:
            type: string
      responses:
        "200":
          description: List of meals with items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MealWithItems"
    post:
      operationId: createMeal
      tags: [meals]
      summary: Log a new meal
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MealInput"
      responses:
        "201":
          description: Meal created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MealWithItems"
  /meals/{id}:
    get:
      operationId: getMeal
      tags: [meals]
      summary: Get a meal by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Meal with items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MealWithItems"
        "404":
          description: Meal not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateMeal
      tags: [meals]
      summary: Update a meal
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MealUpdate"
      responses:
        "200":
          description: Meal updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MealWithItems"
        "404":
          description: Meal not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteMeal
      tags: [meals]
      summary: Delete a meal
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Meal deleted
  /meals/{mealId}/items:
    post:
      operationId: addMealItem
      tags: [meals]
      summary: Add a food item to a meal
      security:
        - bearerAuth: []
      parameters:
        - name: mealId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MealItemInput"
      responses:
        "201":
          description: Meal item added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MealItem"
  /meals/{mealId}/items/{itemId}:
    delete:
      operationId: removeMealItem
      tags: [meals]
      summary: Remove a food item from a meal
      security:
        - bearerAuth: []
      parameters:
        - name: mealId
          in: path
          required: true
          schema:
            type: integer
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Meal item removed

  # ─── Inventory ───────────────────────────────────────────────────────────────
...
  /nutrition/daily:
    get:
      operationId: getDailyNutrition
      tags: [nutrition]
      summary: Get daily nutrition totals
      security:
        - bearerAuth: []
      parameters:
        - name: date
          in: query
          description: Date to calculate for (YYYY-MM-DD). Defaults to today.
          schema:
            type: string
      responses:
        "200":
          description: Daily nutrition summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyNutrition"
  /nutrition/targets:
    get:
      operationId: getNutritionTargets
      tags: [nutrition]
      summary: Get personalized daily nutrition targets based on the user profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Nutrition targets
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NutritionTargets"
        "404":
          description: Profile not found — targets cannot be calculated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── AI ──────────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    # ── Auth ──────────────────────────────────────────────────────────────────
...
    MealItem:
      type: object
      properties:
        id:
          type: integer
        mealId:
          type: integer
        foodId:
          type: ["integer", "null"]
        foodName:
          type: string
        quantityG:
          type: number
        calories:
          type: number
        proteinG:
          type: number
        carbsG:
          type: number
        fatG:
          type: number
        fiberG:
          type: ["number", "null"]
      required:
        - id
        - mealId
        - foodName
        - quantityG
        - calories
        - proteinG
        - carbsG
        - fatG
    MealWithItems:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        mealType:
          type: string
          enum: [breakfast, lunch, dinner, snack]
        date:
          type: string
          format: date
        notes:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/MealItem"
        totalCalories:
          type: number
        totalProteinG:
          type: number
        totalCarbsG:
          type: number
        totalFatG:
          type: number
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - name
        - mealType
        - date
        - items
        - totalCalories
        - totalProteinG
        - totalCarbsG
        - totalFatG
        - createdAt
    MealInput:
      type: object
      properties:
        name:
          type: string
        mealType:
          type: string
          enum: [breakfast, lunch, dinner, snack]
        date:
          type: string
          format: date
        notes:
          type: string
      required:
        - name
        - mealType
        - date
    MealUpdate:
      type: object
      properties:
        name:
          type: string
        mealType:
          type: string
          enum: [breakfast, lunch, dinner, snack]
        date:
          type: string
          format: date
        notes:
          type: string
    MealItemInput:
      type: object
      properties:
        foodId:
          type: integer
        foodName:
          type: string
        quantityG:
          type: number
        caloriesPer100g:
          type: number
        proteinPer100g:
          type: number
        carbsPer100g:
          type: number
        fatPer100g:
          type: number
        fiberPer100g:
          type: number
      required:
        - foodName
        - quantityG
        - caloriesPer100g
        - proteinPer100g
        - carbsPer100g
        - fatPer100g

    # ── Inventory ─────────────────────────────────────────────────────────────
...
    DailyNutrition:
      type: object
      properties:
        date:
          type: string
          format: date
        totalCalories:
          type: number
        totalProteinG:
          type: number
        totalCarbsG:
          type: number
        totalFatG:
          type: number
        totalFiberG:
          type: number
        mealCount:
          type: integer
        meals:
          type: array
          items:
            $ref: "#/components/schemas/MealWithItems"
      required:
        - date
        - totalCalories
        - totalProteinG
        - totalCarbsG
        - totalFatG
        - totalFiberG
        - mealCount
        - meals
    NutritionTargets:
      type: object
      properties:
        dailyCalories:
          type: number
        proteinG:
          type: number
        carbsG:
          type: number
        fatG:
          type: number
        fiberG:
          type: number
        bmr:
          type: number
        tdee:
          type: number
      required:
        - dailyCalories
        - proteinG
        - carbsG
        - fatG
        - fiberG
        - bmr
        - tdee

    # ── AI ────────────────────────────────────────────────────────────────────
...