# lib/api-zod/src/generated/types/mealWithItems.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * NutriAI — AI Nutrition Assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { MealItem } from './mealItem';
import type { MealWithItemsMealType } from './mealWithItemsMealType';

export interface MealWithItems {
  id: number;
  userId: number;
  name: string;
  mealType: MealWithItemsMealType;
  date: Date;
  /** @nullable */
  notes?: string | null;
  items: MealItem[];
  totalCalories: number;
  totalProteinG: number;
  totalCarbsG: number;
  totalFatG: number;
  createdAt: Date;
}